+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * 【extra】        添加MyNLP支持（issue#519@Github）
 * 【json】         添加自定义序列化反序列化支持（issue#I1052A@Gitee）
 * 【dfa】          优化特殊字符构建，优化查找，改为使用StrBuilder
+* 【core】        ZipUtil增加FileFilter参数的重载，支持文件过滤（issue#I11RTP@Gitee）
 
 ### Bug修复
 * 【core】         修复NetUtil.getUsableLocalPort问题（pr#69@Gitee）